bug-hunt #753: fat32 - fix cluster_to_lba uint32 overflow for large clusters
The cluster_to_lba() function computed (cluster - 2) * spc using uint32_t
arithmetic, which silently overflows for valid FAT32 cluster numbers near
0x0FFFFFF0 with sectors-per-cluster >= 17 (e.g., spc=64 with a high cluster
produces ~17B, far exceeding UINT32_MAX).

Fix by promoting the multiplication to uint64_t before truncating back to
uint32_t for the return value.

+++ src/fs/fat32.c
@@ -283,7 +283,7 @@ static void fat_free_chain(uint32_t cluster) {
 
 /* Cluster number → LBA of first sector */
 static uint32_t cluster_to_lba(uint32_t cluster) {
-    return data_start + (cluster - 2) * spc;
+    return data_start + (uint32_t)((uint64_t)(cluster - 2) * (uint64_t)spc);
 }
 
 /* ── 8.3 name comparison (case-insensitive) ─────────────────────────────────── */
